Fix / Add / Update: 设置界面重构、浅色配色对比度与输入栏渲染
Fix:
- 浅色模式 12 套配色预设对比度不足。lightScheme 原先用 primary = seed
  搭 onPrimary = seed.darken(0.6f)，用户气泡对比度只有 1.89~3.90:1，全部
  低于 WCAG 正文 4.5:1（单色预设几乎不可读）。改为 primary/secondary/
  tertiary 先 darken(0.35f) 再配 lighten(0.95f) 的近白前景，12 套全部达标
  （5.51~12.19:1）。仅改 lightScheme，深色方案数值不变
- 输入栏底部出现一条亮带（仅浅色模式，原项目既有）。Compose 把投影和填充
  放在同一 RenderNode，内容不透明时才会跳过被遮挡的阴影；此处填充为 0.85
  透明度，投影渗入内部把大部分区域压暗，未被覆盖的一条反而是唯一正确的
  颜色。将 shadowElevation 从 16dp 改为 0dp
- 输入栏改用 BasicTextField，去掉 M3 TextField 装饰盒按状态分档的容器色，
  避免半透明容器上露底
- 从隐私政策页返回会弹回设置根页而非「关于」。政策页由 ChatScreen 的顶层
  Page.PRIVACY 下移为 SettingsPage.POLICY，导航子树改由 SettingsScreen 持有

Add:
- 设置界面二级导航：入口列表 + 提供商 / 对话 / 外观 / 语音 / 关于五个子页，
  每个入口显示当前状态摘要（模型名、主题、发声开关、版本号）
- ErrorBubble：对话流末尾的错误卡片，取代原错误 Snackbar。errorContainer
  底色、右上角关闭、右下角重试，跟随气泡透明度设置。错误为瞬态 UI 状态，
  不进 ChatMessage 也不写会话历史
- AppInfoSection：应用信息并入关于页；UpdateSection 增加手动检查更新
  （按钮、进度、结果文案与发布页链接）
- SettingsViewModel 增加手动检测更新能力（checkForUpdate /
  dismissUpdateMessage）。启动静默检测仍留在 ChatViewModel
- 21 个 MDI 图标（Apache 2.0）转成 VectorDrawable，共约 10KB；关于页附署名
- tools/svg2vd.py + tools/icons.txt：Iconify → VectorDrawable 批量生成
- tools/contrast_check.py：校验 12 套预设的 WCAG 对比度，未达标返回非 0

Update:
- SettingsScreen.kt 由 1494 行单文件拆为 ui/screen/settings/ 八个文件，
  Section 内部实现未改动，仅按 Kotlin 文件级 private 语义提为 internal
- SettingsGroup 改为 SettingsCard（组标题可选）
- Page 枚举由 CHAT/SETTINGS/PRIVACY 简化为 CHAT/SETTINGS；自动滚底同时
  监听 state.error，使错误卡片出现时滚入视野
- 移除主界面「更多」菜单的「关于」项及 OverlayMenu.onAbout 参数
- 清理死代码：AboutDialog（162 行）、PrivacyPolicyScreen.kt、
  ChatEvent.CheckForUpdate / DismissAboutUpdateMessage、ChatUiState 的三个
  about 字段、LIVE2D_MODEL_SOURCE_URL；ChatScreen.kt 由 646 行降至 474 行
- 精简冗余注释，修正提及已删除 PRIVACY 页与错误 Snackbar 的过期描述
