--- app/src/main/java/com/meapet/mobile/ui/screen/PrivacyPolicyScreen.kt
@@ -1,67 +1,0 @@
-package com.meapet.mobile.ui.screen
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.material3.TopAppBarDefaults
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import com.meapet.mobile.ui.component.PrivacyPolicyContent
-
-/**
- * 隐私政策查看页面。
- *
- * 展示友盟统计 SDK 数据采集相关的隐私政策内容，
- * 用户可从设置页底部跳转至此查看。
- *
- * 内容由 [PrivacyPolicyContent] 统一提供，与首次启动弹窗共享同一份文本。
- */
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun PrivacyPolicyScreen(
-    onBack: () -> Unit
-) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text("隐私政策") },
-                navigationIcon = {
-                    IconButton(onClick = onBack) {
-                        Icon(
-                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
-                            contentDescription = "返回"
-                        )
-                    }
-                },
-                colors = TopAppBarDefaults.topAppBarColors(
-                    containerColor = MaterialTheme.colorScheme.surface.copy(alpha = 0.85f)
-                )
-            )
-        }
-    ) { padding ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(padding)
-                .verticalScroll(rememberScrollState())
-                .padding(horizontal = 16.dp)
-        ) {
-            PrivacyPolicyContent()
-            Spacer(Modifier.height(32.dp))
-        }
-    }
-}
