+++ app/src/main/res/drawable/ic_chevron_right.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24"
+    android:autoMirrored="true">
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M8.59 16.58L13.17 12L8.59 7.41L10 6l6 6l-6 6z" />
+</vector>
